# Optimize Rebind performance and allocations