Avoids many append calls and a string->[]byte conversion.

Benchstat:

```
name            old time/op    new time/op    delta
Rebind-8          1.01µs ± 1%    0.59µs ± 1%  -41.75%  (p=0.000 n=10+10)
RebindBuffer-8    2.55µs ± 0%    2.55µs ± 1%     ~     (p=0.589 n=9+10)

name            old alloc/op   new alloc/op   delta
Rebind-8            493B ± 0%      336B ± 0%  -31.85%  (p=0.000 n=10+10)
RebindBuffer-8      829B ± 0%      829B ± 0%     ~     (all equal)

name            old allocs/op  new allocs/op  delta
Rebind-8            19.0 ± 0%       4.0 ± 0%  -78.95%  (p=0.000 n=10+10)
RebindBuffer-8      21.0 ± 0%      21.0 ± 0%     ~     (all equal)
```